refactor: 리뷰 조회와 삭제 기능을 하나의 훅에서 처리하도록 코드 수정
